requirements.txt
Flask
flask-cors
flask-sqlalchemy
python-dotenv
psycopg2-binary
python-dotenv